% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstockwatch.R
\name{profit_viz}
\alias{profit_viz}
\title{Visualizes trend of a stock price change against the market benchmark within a given period of time}
\usage{
profit_viz(stock_ticker, start_date, end_date, benchmark_ticker)
}
\description{
Visualizes trend of a stock price change against the market benchmark within a given period of time
}
