% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstockwatch.R
\name{volume_change}
\alias{volume_change}
\title{Calculates the daily trading volume change status of a stock within a given period of time}
\usage{
volume_change(stock_ticker, start_date, end_date)
}
\description{
Calculates the daily trading volume change status of a stock within a given period of time
}
