% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstockwatch.R
\name{volume_viz}
\alias{volume_viz}
\title{Visualizes the daily trading volume of a stock using bar plot within a given period of time}
\usage{
volume_viz(stock_ticker, start_date, end_date)
}
\arguments{
\item{stock_ticker}{A string related to ticker of the stock or ETF}

\item{start_date}{A date in string format of "YYYY-MM-DD" related to start of data extraction}

\item{end_date}{A date in string format of "YYYY-MM-DD" related to end of data extraction}
}
\value{
A bar plot with trading volumes
}
\description{
Visualizes the daily trading volume of a stock using bar plot within a given period of time
}
