% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstockwatch.R
\name{percent_change}
\alias{percent_change}
\title{Calculates daily percentage change of a stock price within a given period of time}
\usage{
percent_change(stock_ticker, start_date, end_date)
}
\description{
Calculates daily percentage change of a stock price within a given period of time
}
